requirements.txt
googlesearch-python
django
djangorestframework
ipython
